refactor(cambricon): split cambaricon operators into a separate file
